RF: Minimal CoreGitRepo class
This is merely a first step, moving essential functionality to a new
place, slightly changing its setup. It is by no means comprehensive
or final.
